+++ packages/dashboard/src/api/hooks/statistics/index.ts
@@ -1,0 +1,7 @@
+import { useItemStats } from './useItemStats';
+import { useStockCounts } from './useStockCounts';
+
+export const Statistics = {
+  useItemStats,
+  useStockCounts,
+};
