+++ packages/dashboard/src/api/hooks/utils/index.ts
@@ -1,0 +1,5 @@
+import { useDashboardApi } from './useDashboardApi';
+
+export const Utils = {
+  useDashboardApi,
+};
